fix scanning start of line on Firefox
On many sites, for example Twitter (posts), Yomichan doesn't match any
text from the beginning of the line unless you move the mouse cursor to
the top left corner of the word that starts from the beginning of the
line.

Apparently there were some problems related to the imposter and there
was commit 5fea161 adding a fix for it, but this seemed to cause
problems and 06ba2dc was added to fix that. It didn't handle all of the
cases, though, so this commit is letting everything through on Firefox.